[pluto/form] - refactored form files
