+++ pluto/src/form/external.ts
@@ -7,6 +7,9 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
+export * from "@/form/Context";
 export * from "@/form/Field";
-export * from "@/form/use";
+export * from "@/form/Form";
 export * from "@/form/synced";
+export * from "@/form/use";
+export * from "@/form/useField";
